Revert "moved trefleAPI  and apiMethods to plants folder"
This reverts commit 55add35c944ed2127553ba68e732ce68230fa16e.
